feat: get attachment list for real api
